Use u128 rather than u64 for accoutn and sequence
I agree that u64 is probably large enough for these values, but I also
don't see a reason not to use the largest easily portable integer, which
is u128 in rust since it doesn't come with the same drawbacks as the u256
implemetnation we have.

+++ src/stdsignmsg.rs
@@ -14,9 +14,9 @@ pub struct StdSignMsg {
     /// Chain ID. Example value: "testing"
     pub chain_id: String,
     /// Account number. Example value: 1
-    pub account_number: u64,
+    pub account_number: u128,
     /// Sequence number starts with 0 and should always increase
-    pub sequence: u64,
+    pub sequence: u128,
     /// Fee. Amount should be `None`, and `gas` is the actual gas price.
     pub fee: StdFee,
     /// A list of messages
